requirements.txt
kafka-python-ng==2.2.2
python-dotenv==1.0.0
requests==2.31.0
psycopg2-binary==2.9.9
boto3==1.34.0
